begin popover for UI launch select
