+++ frontend/projects/ui/src/app/pages/apps-routes/app-list/app-list-pkg/app-list-pkg.component.scss
@@ -1,0 +1,3 @@
+ion-popover {
+  --min-width: 300px;
+}
