+++ frontend/projects/ui/src/app/pipes/ui/ui.module.ts
@@ -1,8 +1,8 @@
 import { NgModule } from '@angular/core'
-import { UiPipe } from './ui.pipe'
+import { UiPipe, UiAddressesPipe, AddressTypePipe } from './ui.pipe'
 
 @NgModule({
-  declarations: [UiPipe],
-  exports: [UiPipe],
+  declarations: [UiPipe, UiAddressesPipe, AddressTypePipe],
+  exports: [UiPipe, UiAddressesPipe, AddressTypePipe],
 })
 export class UiPipeModule {}
